Improve security

+++ src/CommunityPatchLauncher/Tasks/UnzipAndExtractCommunityPatchTask.cs
@@ -112,11 +112,16 @@ namespace CommunityPatchLauncher.Tasks
                     string line = string.Empty;
                     while ((line = fileReader.ReadLine()) != null)
                     {
-                        //TODO: IMPORTANT: Check if this is enough protection so that nobody can escape the game folder!
-                        line = line.Replace("../", "");
-                        line = line.Replace("..\\", "");
                         string gameFolder = settingManager.GetValue<string>("GameFolder");
                         string fullPath = gameFolder + line;
+                        fullPath = Path.GetFullPath(fullPath);
+                        FileInfo fileInfo = new FileInfo(fullPath);
+
+                        if (fileInfo.DirectoryName != gameFolder)
+                        {
+                            continue;
+                        }
+
                         if (File.Exists(fullPath))
                         {
                             File.Delete(fullPath);
